server/requirements.txt
# Web framework & server
fastapi==0.128.0
uvicorn[standard]==0.32.1

# Data validation
pydantic>=2.11.1,<3.0.0

#langchain
langchain==1.2.3
langchain-core>=1.2.1,<2.0.0
langchain-community==0.4.1
langchain-text-splitters>=1.0.0

# Vector DB integrations
pinecone==6.0.0
qdrant-client==1.16.2

# Search/tool
tavily-python==0.5.0
firecrawl-py

python-multipart
pymupdf
python-dotenv~=1.2.1

google-genai